msm: mdss: avoid overlay queue for a dirty pipe
During pipe setup, if a failure happens, the pipe
is marked as invalid and an error is returned to
the caller; it is the caller responsibility to honor
the failure and avoid an overlay play in the invalid
configuration. Anyways in order to prevent
configuration problems that can lead to underruns
due a wrong sequence by the caller, add a check
to print a warning and return an error if an
overlay play with an invalid pipe is called.

Change-Id: Ib2497b20eb4117bca8a94c30d3ef3b87ddb79fde
Signed-off-by: Ingrid Gallardo <ingridg@codeaurora.org>